Update README.md
Added a note about the existing Alchemy language converter. Per https://developer.ibm.com/answers/questions/336114/help-integrating-alchemyapi-with-wex.html people are still using this tutorial. It seems like the purpose is to teach someone how to integrate WEX and Bluemix but, in the case of alchemy language, we already have that converter available.